requirements.txt
absl-py==0.9.0
alabaster==0.7.12
aniso8601==8.0.0
argon2-cffi @ file:///C:/ci/argon2-cffi_1596828549974/work
astor==0.8.0
attr==0.3.1
attrs==19.3.0
Babel==2.8.0
backcall==0.2.0
bayesian-optimization==1.2.0
bleach==3.1.5
blinker==1.4
boto3 @ file:///tmp/build/80754af9/boto3_1596840659104/work
botocore @ file:///tmp/build/80754af9/botocore_1596824275563/work
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1596822027882/work
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1594141588948/work
colorama==0.4.3
coremltools @ git+https://github.com/apple/coremltools@9074033980e26657a3e94dfe6dcc5efb235f31b4
cryptography==2.9.2
cycler==0.10.0
Cython @ file:///D:/bld/cython_1594253747180/work
cytoolz==0.10.1
dask @ file:///tmp/build/80754af9/dask-core_1596748257051/work
decorator==4.4.2
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
fire==0.3.1
Flask==1.1.2
Flask-RESTful==0.3.8
future==0.18.2
gast==0.2.2
google-auth @ file:///tmp/build/80754af9/google-auth_1594357566944/work
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.27.2
gym==0.17.2
h5py==2.10.0
idna @ file:///tmp/build/80754af9/idna_1593446292537/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
imagesize==1.2.0
imantics==0.1.12
imgaug==0.4.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1593446525189/work
ipykernel @ file:///C:/ci/ipykernel_1596208728219/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///C:/ci/ipython_1593433781269/work
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi @ file:///C:/ci/jedi_1596472767194/work
Jinja2==2.11.2
jmespath @ file:///tmp/build/80754af9/jmespath_1594304593830/work
joblib @ file:///tmp/build/80754af9/joblib_1594236160679/work
jsonschema @ file:///C:/ci/jsonschema_1594363671836/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1594826976318/work
jupyter-console==6.1.0
jupyter-core==4.6.3
kaggle==1.5.6
Keras==2.3.1
Keras-Applications @ file:///tmp/build/80754af9/keras-applications_1594366238411/work
Keras-Preprocessing==1.1.0
keras2onnx==1.7.0
kiwisolver==1.2.0
labelme2coco==0.1.2
lxml @ file:///C:/ci/lxml_1594826938446/work
Markdown==3.1.1
MarkupSafe @ file:///C:/ci/markupsafe_1594405949945/work
matplotlib @ file:///C:/ci/matplotlib-base_1592846084747/work
mistune @ file:///C:/ci/mistune_1594373272338/work
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
mpmath==1.1.0
nbconvert @ file:///C:/ci/nbconvert_1594372737468/work
nbformat==5.0.7
networkx @ file:///tmp/build/80754af9/networkx_1594377231366/work
notebook @ file:///C:/ci/notebook_1596837179121/work
numpy @ file:///C:/ci/numpy_and_numpy_base_1596233945180/work
oauthlib==3.1.0
olefile==0.46
onnx==1.5.0
onnxconverter-common==1.7.0
onnxmltools==1.5.1
opencv-python==4.3.0.38
opt-einsum==3.1.0
packaging==20.4
pandas @ file:///C:/ci/pandas_1596824386248/work
pandas-datareader @ file:///tmp/build/80754af9/pandas-datareader_1595379313840/work
pandocfilters==1.4.2
parso==0.7.0
pickleshare @ file:///C:/ci/pickleshare_1594374056827/work
Pillow @ file:///C:/ci/pillow_1594298234712/work
pixellib==0.3.0
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycocotools @ git+https://github.com/philferriere/cocoapi.git@2929bd2ef6b451054755dfd7ceb09278f935f7ad#subdirectory=PythonAPI
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyglet==1.5.0
Pygments==2.6.1
PyJWT==1.7.1
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1594392929924/work
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==4.19.18
PyQtWebEngine==5.12.1
pyreadline==2.1
pyrsistent==0.16.0
PySocks @ file:///C:/ci/pysocks_1594394709107/work
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole @ file:///tmp/build/80754af9/qtconsole_1592848611704/work
QtPy==1.9.0
requests @ file:///tmp/build/80754af9/requests_1592841827918/work
requests-oauthlib==1.3.0
rsa==4.0
s3transfer @ file:///C:/ci/s3transfer_1594411925509/work
scikit-image==0.16.2
scikit-learn @ file:///C:/ci/scikit-learn_1592847564598/work
scipy @ file:///C:/ci/scipy_1592916958183/work
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
skl2onnx==1.7.0
snowballstemmer==2.0.0
Sphinx @ file:///home/conda/feedstock_root/build_artifacts/sphinx_1597405755328/work
sphinx-rtd-theme @ file:///home/conda/feedstock_root/build_artifacts/sphinx_rtd_theme_1592482403162/work
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sympy==1.6.2
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
tf2onnx==1.5.3
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toolz==0.10.0
tornado==6.0.4
tqdm @ file:///tmp/build/80754af9/tqdm_1596810128862/work
traitlets==4.3.3
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.25.9
waitress==1.4.4
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
winmltools==1.5.1
wrapt==1.12.1
zipp==3.1.0
